Moved MockExceptionHandler to SIL.XForge
