+++ test/SIL.XForge.Scripture.Tests/Services/MachineBuildServiceTests.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using NSubstitute;
 using NUnit.Framework;
 using SIL.Machine.WebApi;
+using SIL.XForge.Realtime;
 using SIL.XForge.Services;
 
 namespace SIL.XForge.Scripture.Services;
@@ -413,7 +414,7 @@ public class MachineBuildServiceTests
     {
         public TestEnvironment(HttpClient? httpClient = default)
         {
-            var exceptionHandler = new MockExceptionHandler();
+            var exceptionHandler = new MemoryExceptionHandler();
             var httpClientFactory = Substitute.For<IHttpClientFactory>();
             httpClientFactory.CreateClient(Arg.Any<string>()).Returns(httpClient);
